Delete SliderList component and merge its content into SliderTrack
